finish refactor of swapStates
